# openapi.yaml
openapi: 3.0.3
info:
  title: API de Agendamento
  version: 1.0.0
  description: |
    API REST de calendário único com armazenamento em memória. Agendamentos
    ativos não podem compartilhar intervalos sobrepostos. O cancelamento
    exige antecedência configurável (24 horas por padrão).
servers:
  - url: http://localhost:3000
    description: Execução local
paths:
...
  /appointments:
    post:
      summary: Cria um agendamento
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointment'
      responses:
        '201':
          description: Agendamento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Conflito de horário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    get:
      summary: Lista agendamentos
      operationId: listAppointments
      parameters:
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
      responses:
        '200':
          description: Lista ordenada por início
          content:
            application/json:
              schema:
                type: object
                required: [data, total]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Appointment'
                  total:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
...
    Status:
      name: status
      in: query
      schema:
        type: string
        enum: [scheduled, cancelled]
    From:
      name: from
      in: query
      description: Retorna itens cujo fim é posterior a esta data ISO 8601
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      description: Retorna itens cujo início é anterior a esta data ISO 8601
      schema:
        type: string
        format: date-time
  schemas:
    CreateAppointment:
      type: object
      additionalProperties: false
      required: [customerName, customerEmail, startAt, endAt]
      properties:
        customerName:
          type: string
          minLength: 1
          maxLength: 120
        customerEmail:
          type: string
          format: email
          maxLength: 254
        startAt:
          type: string
          format: date-time
          description: ISO 8601 com timezone explícito
        endAt:
          type: string
          format: date-time
        notes:
          type: string
          maxLength: 500
    Appointment:
      allOf:
        - $ref: '#/components/schemas/CreateAppointment'
        - type: object
          required: [id, status, createdAt]
          properties:
            id:
              type: string
              format: uuid
            status:
              type: string
              enum: [scheduled, cancelled]
            createdAt:
              type: string
              format: date-time
            cancelledAt:
              type: string
              format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Regra de negócio ou validação violada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'